+++ examples/use-hbs/mail/index.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import { WelcomeTemplate } from './templates/Welcome';
+
+export default function App() {
+  return {
+    Welcome: {
+      componentFunction: WelcomeTemplate,
+      props: {
+        name: 'string',
+      },
+    },
+  };
+}