+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @jsx-mail/cli
 
+## 1.2.1
+
+### Patch Changes
+
+- Adding specific versions to packages
+- Updated dependencies
+  - @jsx-mail/core@1.0.6
+  - jsx-mail@1.0.4
+
+## 1.2.0
+
+### Minor Changes
+
+- add feature of use hanldebars
+
 ## 1.1.1
 
 ### Patch Changes