+++ packages/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @jsx-mail/core
 
+## 1.0.6
+
+### Patch Changes
+
+- Adding specific versions to packages
+
 ## 1.0.5
 
 ### Patch Changes